% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc24-package.R
\docType{package}
\name{aoc24-package}
\alias{aoc24}
\alias{aoc24-package}
\title{aoc24: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
